feat: add simple search by name of tables
