+++ src/app/index.js
@@ -6,6 +6,7 @@ import Algorithms from '../pages/feature-pages/algorithms'
 import Calculator from '../pages/feature-pages/calculator'
 import Dictionary from '../pages/feature-pages/dictionary'
 import Tables from '../pages/feature-pages/tables'
+import WorkWithTable from '../pages/feature-pages/tables/WorkWithTable'
 
 function App() {
 	return (
@@ -17,6 +18,7 @@ function App() {
 				<Route path='/dictionary' component={Dictionary} />
 				<Route path='/calculator' component={Calculator} />
 				<Route path='/tables' component={Tables} />
+				<Route path='/WorkWithTable' component={WorkWithTable} />
 			</Switch>
 		</div>
 	)
