+++ src/pages/feature-pages/tables/Шкала оценки вероятности ТЭЛА (Revised Geneva Score).json
@@ -69,5 +69,8 @@
         "24": "высокая",
         "25": "высокая"
     }
-]
+],
+"note": "Справочные сведения о таблице. тыры-пыры...",
+"type_of_table": 1,
+"type_of_res": 2
 }
